requirements.txt
sqlalchemy
fastapi
uvicorn
psycopg2
databases
asyncpg
aioredis
pytest
requests
httpx
trio
passlib[bcrypt]
python-jose[cryptography]
python-multipart
web3
py-solc-x
eth_tester
py-evm